requirements.txt
fastapi
uvicorn
jinja2
kafka-python
aiohttp
python-multipart
